rename to briefing_info.py
