+++ src/wblib/figures/internal/sfc_wind_quivers.py
@@ -12,7 +12,7 @@ import seaborn as sns
 import healpy as hp
 import xarray as xr
 
-from wblib.figures.hifs import get_latest_forecast_issue_time
+from wblib.figures.briefing_info import get_latest_forecast_issue_time
 
 CATALOG_URL = "https://tcodata.mpimet.mpg.de/internal.yaml"
 FORECAST_PUBLISH_LAG = "6h"
@@ -32,7 +32,7 @@ def sfc_wind_quivers(briefing_time: pd.Timestamp, lead_hours: str) -> Figure:
     cat = intake.open_catalog(CATALOG_URL)
     ds = cat['HIFS'](refdate=refdate).to_dask().pipe(egh.attach_coords)
     lon_min, lon_max, lat_min, lat_max = FIGURE_BOUNDARIES
-   
+
     windspeed_10m = np.sqrt(ds['10u']**2 + ds['10v']**2)
     lon1 = np.linspace(lon_min, lon_max, MESH_GRID_SIZE)
     lat1 = np.linspace(lat_min, lat_max, MESH_GRID_SIZE)
@@ -55,14 +55,14 @@ def sfc_wind_quivers(briefing_time: pd.Timestamp, lead_hours: str) -> Figure:
         cmap=SPEED_COLORMAP,
         vmin=SPEED_MIN,
         vmax=SPEED_MAX,
-    )   
+    )
     # quiver plot
     Q0 = ax.quiver(lon1[::QUIVER_SKIP], lat1[::QUIVER_SKIP], ds['10u'].sel(time=valid_time).isel(cell = pix)[::QUIVER_SKIP, ::QUIVER_SKIP]
                    , ds['10v'].sel(time = valid_time).isel(cell = pix)[::QUIVER_SKIP, ::QUIVER_SKIP]
-                   , color = 'black', pivot = 'middle', scale_units = 'inches', width = 0.003, scale = 20) 
+                   , color = 'black', pivot = 'middle', scale_units = 'inches', width = 0.003, scale = 20)
     ax.quiverkey(Q0, 0.95, 1.02, 10, r'$10 \frac{m}{s}$', labelpos='E', coordinates='axes', animated=True)
     # 3 m/s wind speed contour
-    im2 = egh.healpix_contour(windspeed_10m.sel(time=valid_time), ax = ax, 
+    im2 = egh.healpix_contour(windspeed_10m.sel(time=valid_time), ax = ax,
             levels = [SPEED_THRESHOLD], colors = 'r'
             )
     # axes formatting
@@ -71,7 +71,7 @@ def sfc_wind_quivers(briefing_time: pd.Timestamp, lead_hours: str) -> Figure:
     _speed_format_axes(briefing_time, issued_time, lead_delta, ax)
     matplotlib.rc_file_defaults()
     return fig
-    
+
 def _speed_format_axes(briefing_time, issued_time, lead_delta, ax):
     lon_min, lon_max, lat_min, lat_max = FIGURE_BOUNDARIES
     valid_time = briefing_time + lead_delta
@@ -82,7 +82,7 @@ def _speed_format_axes(briefing_time, issued_time, lead_delta, ax):
     ax.set_title(title_str)
     ax.coastlines(lw=1.0, color="k")
     ax.set_xticks(np.round(np.linspace(lon_min+5, lon_max-5, 6),0), crs=ccrs.PlateCarree())
-    ax.set_yticks(np.round(np.linspace(lat_min, lat_max, 5),0), crs=ccrs.PlateCarree()) 
+    ax.set_yticks(np.round(np.linspace(lat_min, lat_max, 5),0), crs=ccrs.PlateCarree())
     ax.set_ylabel("Latitude \N{DEGREE SIGN}N")
     ax.set_xlabel("Longitude \N{DEGREE SIGN}E")
     ax.set_xlim([lon_min, lon_max])
@@ -94,5 +94,5 @@ def _speed_format_axes(briefing_time, issued_time, lead_delta, ax):
                 bbox = dict(facecolor='white',
                             edgecolor='none',
                             alpha=1))
-    
+
